Str.ToInt & Str.ToFloat instead of specialized versions
Removes:
- ToInteger
- ToInt64
- ToSingle
- ToDouble

Replaces with just
- ToInt
- ToFloat

+++ Source/simba.nativeinterface.pas
@@ -105,7 +105,7 @@ end;
 
 function TSimbaNativeInterface.WindowHandleFromStr(Str: String): TWindowHandle;
 begin
-  Result := Str.ToInt64(0);
+  Result := Str.ToInt(0);
 end;
 
 procedure TSimbaNativeInterface.PreciseSleep(Milliseconds: UInt32);
